+++ src/components/ThemeList.tsx
@@ -1,4 +1,4 @@
-import { ThemeInfo } from '../../interface';
+import { ThemeInfo } from '../interface';
 import styled from 'styled-components';
 
 import Theme from './Theme';
@@ -26,14 +26,17 @@ const ThemeList = ({ list }: ThemeListProps) => {
 const List = styled.ol`
   display: flex;
   flex-wrap: wrap;
-  gap: 10px;
   min-width: 320px;
   margin-top: 16px;
-  padding: 0 16px;
+  padding: 0 10px;
 
   li {
-    /* width: calc(50% - 10px); */
-    width: 20rem;
+    width: calc(50% - 5px);
+    margin-bottom: 10px;
+
+    :nth-child(odd) {
+      margin-right: 10px;
+    }
   }
 `;
